lark/requirements.txt
Django==4.0
mysql-connector-python==8.0.22
mysqlclient==2.0.2
protobuf==3.19.0
PyMySQL==0.10.1
six==1.15.0
lxml==4.9.3
openpyxl>=3.1.0
beautifulsoup4==4.9.3
urllib3<1.27,>=1.25.4
boto3==1.24.77
requests<3,>=2.31.0
pytz==2021.3
APScheduler==3.8.1
redis==3.5.3
kafka-python==2.0.2
openai==1.84.0
moviepy==1.0.3
tiktoken==0.4.0
python-dotenv==0.19.0
pandas==2.0.0
numpy==1.24.0
httpx==0.24.1
pydantic==2.0.3
cryptography==41.0.7
PyPDF2==3.0.1
sqlalchemy==1.4.31
